requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.24.0
transformers>=4.25.1
gradio>=4.0.0
Pillow>=9.3.0
opencv-python>=4.6.0
tenacity>=8.0.0
huggingface-hub>=0.28.1
tqdm>=4.64.1
accelerate>=0.16.0
safetensors>=0.3.1
sentencepiece>=0.1.96
einops>=0.6.1
matplotlib>=3.6.0
beautifulsoup4>=4.11.1
ftfy>=6.1.1